pyproject.toml
[project]
name = "bloom"
version = "0.1.0"
description = "A concise booking system"
authors = [
    {name = "Nii-Armah",email = "stanleyniiarmah7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
python_files = ["tests.py", "test_*.py"]